Migrate preferences to document client, remove preferences serverless cruft
