+++ notes/cdk.ts
@@ -34,6 +34,7 @@ export class InternoteNotesStack extends InternoteStack {
 
     const table = new dynamo.Table(this, `${id}-table`, {
       tableName: id,
+
       partitionKey: {
         // TODO: One or more parameter values were invalid: Type mismatch for key noteId expected: S actual: M
         name: NOTES_TABLE_PARTITION_KEY,
