--- preferences/db/db.ts
@@ -1,8 +1,0 @@
-import { TypeDynamo } from "type-dynamo";
-
-import { env } from "../env";
-
-export const db = new TypeDynamo({
-  endpoint: env.DYNAMO_ENDPOINT,
-  region: env.SERVICES_REGION,
-});
